Improve switch classification
Adjusted the logic for classifying switches as tactile to look for the
pair of local maxima/minima with the largest difference in force rather
than just the largest local maxima and the smallest local minima
following it. Reduced the force threshold required to classify a switch
as tactile, but added a distance requirement to compensate.

This fixes some tactile switches that were incorrectly classified as
linear.